合并文件问题

+++ src/main/java/MergeFiles.java
@@ -1,4 +1,6 @@
 import java.io.*;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -7,15 +9,15 @@ public class MergeFiles {
 
     public static void handle(List<File> files, String name) {
         Set<String> ipList = removeDup(files);
-        removePrefix(name);
         RecordValidProxy.write(ipList, Constant.filePath + "/" + name);
+        removePrefix(name);
     }
 
     private static void removePrefix(String name) {
         File fileDir = new File(Constant.filePath);
         File[] subFiles = fileDir.listFiles();
         for (File file : subFiles) {
-            if (file.getName().startsWith(name)) {
+            if (file.getName().startsWith(name) && file.getName().length() > 6) {
                 file.delete();
             }
         }
